Move desmume-only test functions to their own file
