+++ tests/conftest.py
@@ -3,70 +3,14 @@ import os
 from pathlib import Path
 import shutil
 import sys
-from typing import Union
 
-from desmume.controls import keymask
 from desmume.emulator import DeSmuME, DeSmuME_SDL_Window
 from ndspy.rom import NintendoDSRom
 import pytest
 
 from patcher.location_types import DigSpotLocation, IslandShopLocation
 from patcher.location_types.island_shop import GD_MODELS
-
-
-class DesmumeEmulator:
-    def __init__(self, py_desmume_instance: tuple[DeSmuME, DeSmuME_SDL_Window]):
-        self.emu = py_desmume_instance[0]
-        self.window = py_desmume_instance[1]
-
-    def open_rom(self, rom_path: str):
-        self.frame = 0
-        self.emu.open(rom_path)
-        self._next_frame()
-
-    def _next_frame(self):
-        self.emu.cycle()
-        self.frame += 1
-        if self.window is not None:
-            self.window.draw()
-            self.window.process_input()
-
-    def wait(self, frames: int):
-        """Idle the emulator for `frames` frames."""
-        starting_frame = self.frame
-        for _ in range(starting_frame, starting_frame + frames):
-            self._next_frame()
-
-    def button_input(self, buttons: Union[int, list[int]], frames: int = 1):
-        """
-        Press buttons.
-
-        Params:
-            buttons: A single button (int) to press, or a list of buttons to simultaneously press.
-            frames: Optional number of frames to hold button for.
-        """
-        if isinstance(buttons, int):
-            buttons = [buttons]
-        for button in buttons:
-            self.emu.input.keypad_add_key(keymask(button))
-        self.wait(frames + 1)
-        for button in buttons:
-            self.emu.input.keypad_rm_key(keymask(button))
-        self.wait(2)
-
-    def touch_input(self, position: tuple[int, int], frames: int = 1):
-        """
-        Touch screen at a given location.
-
-        Params:
-            position: tuple in the form of (x, y) representing the location to touch the screen.
-            frames: Optional number of frames to hold touch screen for.
-        """
-        x, y = position
-        self._next_frame()
-        self.emu.input.touch_set_pos(x, y)
-        self.wait(frames + 1)
-        self.emu.input.touch_release()
+from tests.desmume_utils import DesmumeEmulator
 
 
 @pytest.fixture(scope="session")
